\name{mayo}
\alias{mayo}
\title{Mayo Marker data}
\description{Two marker values with event time and censoring status for
  the subjects in Mayo PBC data}

\format{A data frame with 312 observations and 4 variables: time (event
  time/censoring time), censor (censoring indicator), 
mayoscore4, mayoscore5. The two scores are derived from 4 and 5
covariates respectively.}

\references{Heagerty, P.J., Zheng, Y.  (2005)
  Survival Model Predictive Accuracy and ROC Curves \emph{Biometrics}, \bold{61}, 92 -- 105
} 
\author{Patrick J. Heagerty }
\keyword{survival}
